Added Precision/Recall as Evaluation Metric
Based on Abhishek Takur (4X Kaggle GrandMaster)'s post : 
https://www.linkedin.com/pulse/approaching-almost-any-machine-learning-problem-abhishek-thakur/

Evaluation Metrics
For any kind of machine learning problem, we must know how we are going to evaluate our results, or what the evaluation metric or objective is. For example in case of a skewed binary classification problem we generally choose area under the receiver operating characteristic curve (ROC AUC or simply AUC). In case of multi-label or multi-class classification problems, we generally choose categorical cross-entropy or multiclass log loss and mean squared error in case of regression problems.

And, Based on Following Link :
https://towardsdatascience.com/selecting-the-right-metric-for-skewed-classification-problems-6e0a4a6167a7

If your distribution is skewed then accuracy and AUC-ROC are not good. Its better to use Precision/Recall or some derived metric.